refactor: ContestPage를 pages폴더로 이동
