+++ frontend/src/router.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from 'react-router-dom';
 
-import ContestPage from '@/components/ContestPage';
+import ContestPage from '@/pages/ContestPage';
 
 import App from './App';
